requirements.txt
fastapi>=0.100.0,<1.0.0
uvicorn[standard]>=0.23.0
httpx>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
jieba>=0.42.1
openai>=1.0.0
